Using cmempool instead of lmempool
cmempool is safe for multiple readers/writers simultaneously.

A timer is set for each block request and for unknown reason (yet)
hrtimer expires permaturly. A condition is added and the timer is
rescheduled for the remaining time, in case it were to fail.

Remove device: new function to remove virtual block device
by its backing device name